maps, map, and home views

+++ src/app/routes/api.ts
@@ -26,10 +26,15 @@ api.get("/maps", async (req, res) => {
 });
 
 api.post("/maps", requireAuth, async (req, res) => {
-  const { map_name, description } = req.body;
+  const { map_name, description, published } = req.body;
   const user_id = (req.user as User).user_id;
   if (user_id === undefined) throw new Error("User has no user_id");
-  const map = await TwilightMap.create({ user_id, description, map_name });
+  const map = await TwilightMap.create({
+    user_id,
+    description,
+    map_name,
+    published,
+  });
   res.json(map);
 });
 
@@ -43,7 +48,7 @@ api.get("/maps/:id/latest/large.jpg", async (req, res) => {
   const map_id = parseInt(req.params.id);
   const map = await TwilightMap.get(map_id);
   const latest = map.versions[map.versions.length - 1];
-  const jpg = await compositor.drawTTSMap(latest);
+  const jpg = await compositor.drawTTSMap(latest, 1000, "#dde5ea");
   res.type("jpg").send(jpg);
 });
 
@@ -51,16 +56,17 @@ api.get("/maps/:id/latest/small.jpg", async (req, res) => {
   const map_id = parseInt(req.params.id);
   const map = await TwilightMap.get(map_id);
   const latest = map.versions[map.versions.length - 1];
-  const jpg = await compositor.drawTTSMap(latest, 300);
+  const jpg = await compositor.drawTTSMap(latest, 250, "#dde5ea");
   res.type("jpg").send(jpg);
 });
 
 api.put("/maps/:id", requireAuth, async (req, res) => {
   const map_id = parseInt(req.params.id);
   const map = await TwilightMap.get(map_id);
-  const { newVersion, map_name, description, redirect } = req.body;
+  const { newVersion, map_name, published, description, redirect } = req.body;
   if (map_name !== undefined) map.map_name = map_name;
   if (description !== undefined) map.description = description;
+  if (published !== undefined) map.published = published;
   if (newVersion !== undefined && newVersion !== map.latest)
     map.versions.push(newVersion);
   await map.save();
@@ -81,24 +87,34 @@ api.post("/maps/:id/comments", requireAuth, async (req, res) => {
 });
 
 api.put("/maps/:id/comments/:commentId", requireAuth, async (req, res) => {
+  console.log("PUT", req.path);
   const { text, redirect } = req.body;
+  console.log(text, redirect);
   const comment_id = parseInt(req.params.commentId);
   const comment = await MapComment.get(comment_id);
-  comment.modify({ text });
-  await comment.save();
-  res.json(comment);
-  if (redirect !== undefined) res.redirect(redirect);
-  else res.json(comment);
+  const user = req.user as User;
+  if (comment.user_id !== user.user_id) {
+    res.status(403).send("Forbidden");
+  } else {
+    comment.modify({ text });
+    await comment.save();
+    if (redirect !== undefined) res.redirect(redirect);
+    else res.json(comment);
+  }
 });
 
 api.delete("/maps/:id/comments/:commentId", async (req, res) => {
   const { redirect } = req.body;
+  const user = req.user as User;
   const comment_id = parseInt(req.params.commentId);
   const comment = await MapComment.get(comment_id);
-  await comment.delete();
-  res.send(comment);
-  if (redirect !== undefined) res.redirect(redirect);
-  else res.json(comment);
+  if (comment.user_id !== user.user_id) {
+    res.status(403).send("Forbidden");
+  } else {
+    await comment.delete();
+    if (redirect !== undefined) res.redirect(redirect);
+    else res.json(comment);
+  }
 });
 
 export default api;
